Separate ciphertext and arithmetic functions
