+++ include/core/ggsw/glwegadget.h
@@ -2,6 +2,7 @@
 #define PARTIAL_GGSW_H
 
 #include "bivariate_polynomial.h"
+#include "glwe_ciphertext.h"
 #include "glwe_key.h"
 #include "glwegadget_ciphertext.h"
